mypt-ho-job-api/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
Markdown==3.4.1
mysqlclient==2.1.1
numpy==1.23.4
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pycryptodome==3.15.0
pycryptodomex==3.15.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.6
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
